add fixup for image1dbuffer with bgra format on Intel
also fix condition for CL_LUMINANCE and CL_INTENSITY (ref #625)
