+++ src/CMakeLists.txt
@@ -49,6 +49,7 @@ add_library(OpenCL-objects OBJECT
   device.cpp
   device_properties.cpp
   event.cpp
+  image_format.cpp
   init.cpp
   kernel.cpp
   log.cpp
